requirements.txt
crewai>=0.80.0
crewai-tools>=0.14.0
python-dotenv>=1.0.0
requests>=2.31.0
firecrawl-py>=1.0.0
